Move connector and use commands into their own files
And also move their tests to the corresponding test file
